improve sorting of URLs
It turns out that Chrome/WebKit and PHP do not preserve the original order
of array items if they are considered equal by the sort function. As we
want to preserve the original order of the URLs, just put the ones with
titles to the front, we now add the original index to the object itself
and compare the original indexes in case both URL have a title or no title.