introduce useKeywords TypoScript setting
defaults to 0, if it is turned on, it:
* shows a keywords search field in extended search, querying the subjects index
* displays keywords in result details
* makes each of these keywords a link that stars a search for that keyword

In addition, toggling the extended search form has been improved.

Note: in practice this doesn’t seem like a particularly good feature
as keywords are rarely given to records consistently or completely. Neither in
a single catalogue nor across the catalogues of a metasearch.